imovirtual/imovirtual/src/requirements.txt
requests
tqdm
pandas
pyarrow
beautifulsoup4
google-cloud-storage
python-dotenv